Monthy update
Q: Why are sub-domains blocked (red) if it's already covered blocking the top domain?
A: Because some pages loading the sub-domains as resources which then is not blocked. 

Q: Why are cookies blocked for blocked sub-domains and domains if it's already blocked?
A: It's only a visible indicator for me that these domains are 100% safe to block. 

The mentioned redundancies have no impact on loading speed or something. It's just an indicator and cover all possible scenarios.